fix: Replace android.graphics.Rect with Compose Rect to fix preview crashes
- Changed DetectedTextBlock to use androidx.compose.ui.geometry.Rect instead of android.graphics.Rect
- Updated CameraViewModel to convert ML Kit Rect to Compose Rect with float values
- Fixed all preview functions in CameraScreen to use Compose Rect
- This resolves InvocationTargetException in Compose previews caused by Android framework classes

The issue was that Android framework classes like android.graphics.Rect cannot be instantiated in Compose preview environment, causing reflection errors. Compose's own Rect class works in both runtime and preview contexts.

+++ app/src/main/java/com/example/globaltranslation/ui/camera/CameraViewModel.kt
@@ -1,6 +1,7 @@
 package com.example.globaltranslation.ui.camera
 
-import android.graphics.Rect
+import android.graphics.Rect as AndroidRect
+import androidx.compose.ui.geometry.Rect
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.globaltranslation.core.provider.CameraTranslationProvider
@@ -111,10 +112,10 @@ class CameraViewModel @Inject constructor(
                                 originalText = block.originalText,
                                 translatedText = block.translatedText,
                                 boundingBox = Rect(
-                                    block.boundingBox.left,
-                                    block.boundingBox.top,
-                                    block.boundingBox.right,
-                                    block.boundingBox.bottom
+                                    block.boundingBox.left.toFloat(),
+                                    block.boundingBox.top.toFloat(),
+                                    block.boundingBox.right.toFloat(),
+                                    block.boundingBox.bottom.toFloat()
                                 )
                             )
                         }
